refactor away editblogview
