+++ Apps/iOS/Postalgic/Views/Blog List/BlogFormView.swift
@@ -11,10 +11,32 @@ import SwiftUI
 struct BlogFormView: View {
     @Environment(\.modelContext) private var modelContext
     @Environment(\.dismiss) private var dismiss
-
-    @State private var name = ""
-    @State private var url = ""
-
+    
+    // Determines if we're creating a new blog or editing an existing one
+    private var isEditing: Bool
+    
+    // Optional blog for editing mode
+    private var blog: Blog?
+    
+    @State private var name: String
+    @State private var url: String
+    
+    // Initialize for creating a new blog
+    init() {
+        self.isEditing = false
+        self.blog = nil
+        _name = State(initialValue: "")
+        _url = State(initialValue: "")
+    }
+    
+    // Initialize for editing an existing blog
+    init(blog: Blog) {
+        self.isEditing = true
+        self.blog = blog
+        _name = State(initialValue: blog.name)
+        _url = State(initialValue: blog.url)
+    }
+    
     var body: some View {
         NavigationStack {
             Form {
@@ -27,7 +49,7 @@ struct BlogFormView: View {
                         .textContentType(.URL)
                 }
             }
-            .navigationTitle("New Blog")
+            .navigationTitle(isEditing ? "Edit Blog" : "New Blog")
             .toolbar {
                 ToolbarItem(placement: .cancellationAction) {
                     Button("Cancel") {
@@ -36,7 +58,11 @@ struct BlogFormView: View {
                 }
                 ToolbarItem(placement: .confirmationAction) {
                     Button("Save") {
-                        addBlog()
+                        if isEditing {
+                            updateBlog()
+                        } else {
+                            addBlog()
+                        }
                         dismiss()
                     }
                     .disabled(name.isEmpty || url.isEmpty)
@@ -44,14 +70,26 @@ struct BlogFormView: View {
             }
         }
     }
-
+    
     private func addBlog() {
         let newBlog = Blog(name: name, url: url)
         modelContext.insert(newBlog)
+    }
+    
+    private func updateBlog() {
+        if let blogToUpdate = blog {
+            blogToUpdate.name = name
+            blogToUpdate.url = url
+        }
     }
 }
 
 #Preview {
     BlogFormView()
+        .modelContainer(for: [Blog.self], inMemory: true)
+}
+
+#Preview {
+    BlogFormView(blog: Blog(name: "Test Blog", url: "https://example.com"))
         .modelContainer(for: [Blog.self], inMemory: true)
 }
